Move into test srcset
