--- src/test/groovy/groovylicioustest/ClientConfig.groovy
@@ -1,8 +1,0 @@
-package groovylicioustest
-
-import ga.ozli.minecraftmods.groovylicious.transform.Config
-
-@Config
-class ClientConfig {
-    static boolean sup = true
-}
